Clear the shadowing, unsafe-any and useless-undefined lint families and promote them to errors
A typed Buffer alias in the supervisor replaces the `any` global that
@cloudflare/workers-types declares; explicit Uint8Array types at the source
remove the remaining unsafe-any findings; inner variables are renamed instead
of shadowing; Request clones pass a body only for methods that allow one.
The cleared families now lint as errors; complexity and nested ternaries stay
warnings until their refactors land.

+++ packages/supervisor/src/workspace.ts
@@ -43,7 +43,9 @@ export async function executeWorkspace(
       try {
         for (;;) {
           const chunk = await reader.read();
-          buffer += chunk.done ? decoder.decode() : decoder.decode(chunk.value, { stream: true });
+          buffer += chunk.done
+            ? decoder.decode()
+            : decoder.decode(chunk.value as Uint8Array, { stream: true });
           if (buffer.length > 8_000_000) throw new Error("Command stream exceeds its limit");
           for (;;) {
             const boundary = buffer.indexOf("\n");
