+++ tests/codex/codex.test.ts
@@ -13,6 +13,7 @@ import {
   type RuntimeEvent,
   runPromise,
 } from "../../packages/agent-api/src/index.js";
+import { Buffer } from "../../packages/supervisor/src/buffer.js";
 import { CodexJob } from "../../packages/supervisor/src/codex.js";
 
 const cleanup: (() => Promise<unknown>)[] = [];
@@ -33,7 +34,7 @@ it("runs native Codex shell calls in the separate exec-server workspace and rest
   let functionIssued = false;
   const handle = async (request: IncomingMessage, response: ServerResponse) => {
     const chunks: Uint8Array[] = [];
-    for await (const chunk of request) chunks.push(chunk);
+    for await (const chunk of request) chunks.push(chunk as Uint8Array);
     const body = JSON.parse(Buffer.concat(chunks).toString()) as Record<string, unknown>;
     requests.push(body);
     response.writeHead(200, { "content-type": "text/event-stream" });
