+++ tests/harnesses/http.ts
@@ -1,6 +1,8 @@
 import { once } from "node:events";
 import { createServer, type IncomingMessage, type ServerResponse } from "node:http";
 
+import { Buffer } from "../../packages/supervisor/src/buffer.js";
+
 /** Real HTTP boundary, including streaming bodies and disconnect cancellation. */
 export async function serveFetch(fetch: (request: Request) => Promise<Response>) {
   const handle = async (incoming: IncomingMessage, outgoing: ServerResponse) => {
@@ -8,7 +10,7 @@ export async function serveFetch(fetch: (request: Request) => Promise<Response>)
     outgoing.on("close", () => abort.abort());
     try {
       const chunks: Uint8Array[] = [];
-      for await (const chunk of incoming) chunks.push(chunk);
+      for await (const chunk of incoming) chunks.push(chunk as Uint8Array);
       const headers = new Headers();
       for (const [key, value] of Object.entries(incoming.headers))
         if (value) headers.set(key, Array.isArray(value) ? value.join(", ") : value);
